refactor: Centralize icon pack file path retrieval
This commit consolidates the logic for obtaining icon pack file paths into
`GetHomeDataUseCase`. Previously, this responsibility was handled by a
separate `GetGridItemsIconPackInfoFilePathUseCase`.

The `GetGridItemsIconPackInfoFilePathUseCase` has been removed entirely.
The `HomeViewModel` no longer directly collects the icon pack info file
paths. Instead, `GetHomeDataUseCase` now computes and provides these paths
as part of the `HomeData` object.

This change simplifies the architecture by reducing the number of
use cases and centralizing data fetching within `GetHomeDataUseCase`,
making the data flow more straightforward and easier to manage.

The `HomeScreen` and its sub-components will now receive the
`iconPackInfoFilePaths` directly from the `homeData` object.
